skipcastify/pyproject.toml
[project]
name = "skipcastify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "feedgen>=1.0.0",
    "feedparser>=6.0.11",
    "flask>=3.0.3",
    "pyaml>=25.1.0",
    "python-dotenv>=1.0.1",
]
